* prac/1:
  test
  layered architecture pattern comments update
  layered architecture pattern 계층별 코드 수정
  layered architecture pattern 계층별 코드 수정
  layered architecture pattern 계층별 코드 수정
  미사용 middleware 삭제
  console.log 수정

+++ service/comments.service.js
@@ -1,31 +1,38 @@
 const CommentRepository = require('../repositories/Comments.repository');
+const PostRepository = require('../repositories/posts.repository');
 
 class CommentService {
   commentRepository = new CommentRepository();
+  postRepository = new PostRepository();
 
   createComment = async (req, res, next) => {
-    try {
-      const { comment } = req.body;
-      const { postId } = req.params;
-      const { user } = res.locals;
-      const result = {
-        postId,
-        userId: user.userId,
-        nickname: user.nickname,
-        comment,
-      };
+    // 댓글 내용 검증
+    const { comment } = req.body;
+    if (!comment) throw new Error('댓글을 입력해주세요.');
 
-      await this.commentRepository.createComment(result);
+    // 게시글이 있는지 검증
+    const { postId } = req.params;
+    const posts = await this.postRepository.findByPost(postId);
+    if (!posts) throw new Error('존재하지 않는 게시글입니다.');
+    const { user } = res.locals;
 
-      return;
-    } catch (error) {
-      res.status(400).send({ message: error.message });
-    }
+    // repository로 보낼 데이터 가공
+    const result = {
+      postId,
+      userId: user.userId,
+      nickname: user.nickname,
+      comment,
+    };
+
+    await this.commentRepository.createComment(result);
+
+    return;
   };
 
   getComment = async (req, res) => {
     const { postId } = req.params;
-
+    const posts = await this.postRepository.findByPost(postId);
+    if (!posts) throw new Error('존재하지 않는 게시글입니다.');
     const comments = await this.commentRepository.getComment(postId);
 
     return comments;
@@ -38,24 +45,33 @@ class CommentService {
   };
 
   updateComment = async (req, res, next) => {
-    try {
-      const { commentId } = req.params;
-      const { comment } = req.body;
+    const { commentId } = req.params;
+    const { user } = res.locals;
+    const { comment } = req.body;
+    if (!comment) throw new Error('댓글 내용을 입력해주세요.');
 
-      await this.commentRepository.updateComment(commentId, comment);
-    } catch (error) {
-      res.status(400).send({ message: error.message });
-    }
+    // 존재하는 댓글인지 검증
+    const getCommentId = await this.commentRepository.getCommentId(commentId);
+    if (!getCommentId) throw new Error('존재하지 않는 댓글입니다.');
+
+    // 댓글 작성자와 로그인된 유저의 ID를 검증
+    if (user.userId !== getCommentId.userId)
+      throw new Error('로그인된 사용자와 게시자가 다릅니다');
+
+    await this.commentRepository.updateComment(commentId, comment);
   };
 
   destroyComment = async (req, res, next) => {
-    try {
-      const { commentId } = req.params;
+    const { commentId } = req.params;
+    const { user } = res.locals;
+    const getCommentId = await this.commentRepository.getCommentId(commentId);
 
-      await this.commentRepository.destroyComment(commentId);
-    } catch (error) {
-      res.status(400).send({ message: error.message });
-    }
+    if (!getCommentId) throw new Error('존재하지 않는 댓글입니다.');
+
+    if (user.userId !== getCommentId.userId)
+      throw new Error('로그인된 사용자와 게시자가 다릅니다');
+
+    await this.commentRepository.destroyComment(commentId);
   };
 }
